Speed-up Sync process
This is a simple change which speeds up Sync process on systems with a high amount of Snapshots. It gets only Snapshots desired for the targeted FS and reduces output and thus Parsing time of the get command in the first place, and schould greatly reduce later processes (sorting arrays, mathing snaps) due to the reduced overhead.