Add comprehensive installation instructions
Converts the old INSTALL file to a markdown-formatted INSTALL.md file
that contains comprehensive installation instructions for supported
operating systems, and an outline that can be used on others where
Sanoid/Syncoid ought to work.

The Installation section is complete for CentOS/RHEL, and has been
"dressed up" for FreeBSD/Ubuntu. Those ones still need to be completed.
I don't cover the topic of running Syncoid periodically or how to do it
because I don't know. It's probably out of scope for the Installation
guide anyway.

The Configuration sections is intended to walk through some common
setups for sanoid.conf. I suggest just cleaning up the existing conf
file examples and putting them into that section.