Do not require user to be specified for syncoid
This brings the behavior of syncoid into line with the documentation,
which suggests that :, rather than @, is required when specifying a
remote dataset. The behavior of a TARGET or SOURCE that begins with :
is altered to allow for local dataset names with colons in them.

The name of the control socket is changed slightly.