Copy Tweak
I assume this is for sanoid development? Just editing to make it more clear, assuming that's the case (the phrasing stuck me as odd, saying it should *never* be used).

If it's there for another reason, feel free to reject or edit, just wanted to make the docs slightly more clear!