Add --skip-redundant-sync-snap
Commonly we want to sync all files on a filesystem, even newer than the
last snapshot. So by default syncoid creates a sync snapshot and sends
that over. This can take many seconds - even if no data has changed.

This patch adds a new flag, --skip-redundant-sync-snap, which checks the
`written@...` property on the filesystem to see if anything has been
written since the latest snapshot. If not, it avoids creating one. This
avoids transferring anything at all if nothing has changed.

If no sync snapshot was created, pruning of old sync snapshots on the
target is also skipped.

At the moment, it uses the last snapshot (of any kind). It is assumed
that snapshots are never deleted on the target, except via the syncoid
pruning mechanism.