Add the 'getbinarypath' function
STATE:
I am using sanoid with debian 'bullseye/sid',
while trying to sync some ZFS dataset between servers
with a usual user, it seems that some zfs binaries
are not found on the remote server.

FIX:
The new function 'getbinarypath' is defined to get
a full path when using 'zfs' binaries on localhost
as well as on the remotehost.

Signed-off-by: Jeremy MAURO <jeremy.mauro@gmail.com>