fix: Warn when newest snapshot is in the future
Option --monitor-snapshots issues a Warning or a Critical when newest
snapshot is too old based on the thresholds defined in the Sanoid
configuration.

However, it reports an _OK_ status when newest snapshot is in the
future. This can happen when system time is wrong (in the future) for
some time and then returns to normal. In such a situation, this may
prevent the creation of new snapshots for a while and so we need to be
alerted.

Signed-off-by: Nicolas Payart <npayart@gmail.com>