Provide reason in warning message
I got this warning when I wanted to create a template with "recursive = yes" and had to look at the sourcecode to find it's not allowed in templates.
This PR proposes to clarify the reason in the warning message to make it less confusing. Please revise the language as I'm not a native speaker.