add --syncstats option
The `--syncstats` option is to support general per-dataset instrumentation of the syncoid process, and allows you to specify a file path to dump per-dataset stats to.  The specified file will contain json with stats for each dataset that is considered.

The original use case driving this change is to parse the output file after the sync is complete, and send the data to your metric tracking system/TSDB of choice.

For this commit, the only stat currently added is the sync duration (`sync_duration_s`).  However, the code/output has been set up to make it easy to add more per-dataset stats in the future.  For example, `delta_bytes` (the size of the incremental delta in bytes) would be another nice stat to add later.