syncoid: add --filter flag (hack hack hack)
We want to run syncoid with --no-sync-snap but we do not want to
consider all available snapshots as suitable incremental sources,
certain specially named snapshots.  With the --filter flag, we can
specify a regexp that will force syncoid to only "see" matching
snapshots.  This is not ideal: if we ever change which snapshots we
use, filtering in this way will make it impossible to find a common
snapshot or bookmark to use as an incremental source.  But that would
require quite a bit more thought to capture all of the difference
cases.